feat: 알코올 도수를 hook으로 분리
