+++ frontend/src/@types/image.d.ts
