+++ frontend/src/@types/product.ts
