+++ frontend/src/apis/request/member.ts
@@ -1,6 +1,6 @@
 import { axios, axiosWithAccessToken } from '../axios';
 import { API_PATH } from '@/constants/path';
-import { SignupUser } from '@/types/user';
+import { SignupUser } from '@/@types/user';
 
 const requestSignup = (userData: SignupUser) => {
   return axios.post(API_PATH.MEMBER.SIGNUP, userData);
