+++ frontend/src/apis/request/storage.ts
@@ -1,7 +1,7 @@
 import { axiosWithAccessToken } from '../axios';
 
 import { API_PATH } from '@/constants/path';
-import { ImageIdType } from '@/types/id';
+import { ImageIdType } from '@/@types/id';
 
 const requestSaveImage = (image: File) => {
   const formData = new FormData();
