+++ frontend/src/components/Header/Header.tsx
@@ -6,7 +6,7 @@ import Search from './Search/Search';
 
 import { BROWSER_PATH } from '@/constants/path';
 import { useDrinkUpperCategory } from '@/hooks/useCategory';
-import { CategoryType } from '@/types/category';
+import { CategoryType } from '@/@types/category';
 
 import * as S from './Header.styled';
